#33333 Adds the ability to publish alembic/mantra outputs.
If using the tk-houdini-alembicnode to generate alembic caches or the
tk-houdini-mantranode to render images, the publisher will now prompt the
user to publish the files generated by these nodes. The nodes will display
as secondary publishes in the UI and will be automatically checked if files
exist on disk for the current version. Closes #25.