#37684: Addresses a layout quirk in Maya 2017 on Windows.
The quirk resulted in a visible UI artifact in the main publisher dialog after a screenshot was taken. It appears to be a quirk of some kind related to Qt5 and the layout of the app's publish_details_form widget. The fix is just to tweak the height of that widget. Weird.